Restore HIV_PREVIOUS_EXPOSURES migration; add missing setAutoIncrement (#109)
* Revert "revert: UHM-5039: Migrate HIV_PREVIOUS_EXPOSURES table - part 1"

This reverts commit 24035839b830a3f879a5effb75c042c82ecc5087.

This restores commit 18dcb90cba907669ca0991ec931f812080efaaf6:

UHM-5039 Migrate HIV_PREVIOUS_EXPOSURES table - part 1 (#106)
https://pihemr.atlassian.net/browse/UHM-5039

* Add missing setAutoIncrement for previous exposures group table

* Only revert if the necessary tables are present

+++ etl/src/main/java/org/pih/hivmigration/etl/sql/PcrTestsMigrator.groovy
@@ -142,8 +142,10 @@ class PcrTestsMigrator extends SqlMigrator {
 
     @Override
     def void revert() {
-        executeMysql("delete from obs where encounter_id in (select encounter_id from hivmigration_pcr_tests)")
-        executeMysql("delete from encounter where encounter_id in (select encounter_id from hivmigration_pcr_tests)")
-        executeMysql("drop table if exists hivmigration_pcr_tests")
+        if(tableExists("hivmigration_pcr_tests")) {
+            executeMysql("delete from obs where encounter_id in (select encounter_id from hivmigration_pcr_tests)")
+            executeMysql("delete from encounter where encounter_id in (select encounter_id from hivmigration_pcr_tests)")
+            executeMysql("drop table hivmigration_pcr_tests")
+        }
     }
 }
